+++ backend/src/main/java/com/portafolio/mgb/Controller/ProyectosController.java
@@ -38,7 +38,7 @@ public class ProyectosController {
     }
 
     @GetMapping("/list/{id}")
-    public ResponseEntity<List<Proyectos>> list(@PathVariable long id) {
+    public ResponseEntity<List<Proyectos>> list(@PathVariable int id) {
         if (PerServ.buscarPersona(id) == null) {
             return new ResponseEntity("No se ha podido encontrar la persona", HttpStatus.BAD_REQUEST);
         }
