+++ backend/src/main/java/com/portafolio/mgb/Interface/IProyectosService.java
@@ -8,7 +8,7 @@ public interface IProyectosService {
 
     public List<Proyectos> listProyectos();
     
-    public List<Proyectos> listProyectosByIdPersona(long id);
+    public List<Proyectos> listProyectosByIdPersona(int id);
 
     public Proyectos buscarProyecto(int id);
